Add CharacterList Screen (replaces TabOne Screen)
The building of this took place in a couple of stages

First, I worked against dummy data without the fetch, while I got used to the navigation system. Unfortunately, as I was in a rush, I did not commit these in separate steps.
Then, once the data was out and the navigation was in place, I worked on a character profile.
Then, I implemented fetch. Then, I added parameters into the character profile navigation.